TWS API v9.76.01 fixes:  + renamed StdAfx.h/StdAfx.cpp to stdafx.h/stdafx.cpp  + added x64 builds  + fixed warnings (wrong printf-modifiers, type casts, missing returns statements)  + fixed vs projects/solutions  + updated vs projects to vs 2019  + removed dll builds of TwsSocketClient/TestCppClient

+++ samples/Cpp/TestCppClient/TestCppClient.cpp
@@ -1,7 +1,7 @@
 ﻿/* Copyright (C) 2019 Interactive Brokers LLC. All rights reserved. This code is subject to the terms
  * and conditions of the IB API Non-Commercial License or the IB API Commercial License, as applicable. */
 
-#include "StdAfx.h"
+#include "stdafx.h"
 
 #include "TestCppClient.h"
 
@@ -1581,7 +1581,7 @@ void TestCppClient::printContractDetailsMsg(const ContractDetails& contractDetai
 }
 
 void TestCppClient::printContractDetailsSecIdList(const TagValueListSPtr &secIdList) {
-	const int secIdListCount = secIdList.get() ? secIdList->size() : 0;
+	const int secIdListCount = secIdList.get() ? (int)secIdList->size() : 0;
 	if (secIdListCount > 0) {
 		printf("\tSecIdList: {");
 		for (int i = 0; i < secIdListCount; ++i) {
@@ -1839,7 +1839,7 @@ void TestCppClient::securityDefinitionOptionalParameterEnd(int reqId) {
 
 //! [softDollarTiers]
 void TestCppClient::softDollarTiers(int reqId, const std::vector<SoftDollarTier> &tiers) {
-	printf("Soft dollar tiers (%lu):", tiers.size());
+	printf("Soft dollar tiers (%zu):", tiers.size());
 
 	for (unsigned int i = 0; i < tiers.size(); i++) {
 		printf("%s\n", tiers[i].displayName().c_str());
@@ -1849,7 +1849,7 @@ void TestCppClient::softDollarTiers(int reqId, const std::vector<SoftDollarTier>
 
 //! [familyCodes]
 void TestCppClient::familyCodes(const std::vector<FamilyCode> &familyCodes) {
-	printf("Family codes (%lu):\n", familyCodes.size());
+	printf("Family codes (%zu):\n", familyCodes.size());
 
 	for (unsigned int i = 0; i < familyCodes.size(); i++) {
 		printf("Family code [%d] - accountID: %s familyCodeStr: %s\n", i, familyCodes[i].accountID.c_str(), familyCodes[i].familyCodeStr.c_str());
@@ -1859,13 +1859,13 @@ void TestCppClient::familyCodes(const std::vector<FamilyCode> &familyCodes) {
 
 //! [symbolSamples]
 void TestCppClient::symbolSamples(int reqId, const std::vector<ContractDescription> &contractDescriptions) {
-	printf("Symbol Samples (total=%lu) reqId: %d\n", contractDescriptions.size(), reqId);
+	printf("Symbol Samples (total=%zu) reqId: %d\n", contractDescriptions.size(), reqId);
 
 	for (unsigned int i = 0; i < contractDescriptions.size(); i++) {
 		Contract contract = contractDescriptions[i].contract;
 		std::vector<std::string> derivativeSecTypes = contractDescriptions[i].derivativeSecTypes;
 		printf("Contract (%u): %ld %s %s %s %s, ", i, contract.conId, contract.symbol.c_str(), contract.secType.c_str(), contract.primaryExchange.c_str(), contract.currency.c_str());
-		printf("Derivative Sec-types (%lu):", derivativeSecTypes.size());
+		printf("Derivative Sec-types (%zu):", derivativeSecTypes.size());
 		for (unsigned int j = 0; j < derivativeSecTypes.size(); j++) {
 			printf(" %s", derivativeSecTypes[j].c_str());
 		}
@@ -1876,7 +1876,7 @@ void TestCppClient::symbolSamples(int reqId, const std::vector<ContractDescripti
 
 //! [mktDepthExchanges]
 void TestCppClient::mktDepthExchanges(const std::vector<DepthMktDataDescription> &depthMktDataDescriptions) {
-	printf("Mkt Depth Exchanges (%lu):\n", depthMktDataDescriptions.size());
+	printf("Mkt Depth Exchanges (%zu):\n", depthMktDataDescriptions.size());
 
 	for (unsigned int i = 0; i < depthMktDataDescriptions.size(); i++) {
 		printf("Depth Mkt Data Description [%d] - exchange: %s secType: %s listingExch: %s serviceDataType: %s aggGroup: %s\n", i, 
@@ -1897,7 +1897,7 @@ void TestCppClient::tickNews(int tickerId, time_t timeStamp, const std::string& 
 
 //! [smartcomponents]]
 void TestCppClient::smartComponents(int reqId, const SmartComponentsMap& theMap) {
-	printf("Smart components: (%lu):\n", theMap.size());
+	printf("Smart components: (%zu):\n", theMap.size());
 
 	for (SmartComponentsMap::const_iterator i = theMap.begin(); i != theMap.end(); i++) {
 		printf(" bit number: %d exchange: %s exchange letter: %c\n", i->first, std::get<0>(i->second).c_str(), std::get<1>(i->second));
@@ -1915,7 +1915,7 @@ void TestCppClient::tickReqParams(int tickerId, double minTick, const std::strin
 
 //! [newsProviders]
 void TestCppClient::newsProviders(const std::vector<NewsProvider> &newsProviders) {
-	printf("News providers (%lu):\n", newsProviders.size());
+	printf("News providers (%zu):\n", newsProviders.size());
 
 	for (unsigned int i = 0; i < newsProviders.size(); i++) {
 		printf("News provider [%d] - providerCode: %s providerName: %s\n", i, newsProviders[i].providerCode.c_str(), newsProviders[i].providerName.c_str());
@@ -1971,7 +1971,7 @@ void TestCppClient::headTimestamp(int reqId, const std::string& headTimestamp) {
 
 //! [histogramData]
 void TestCppClient::histogramData(int reqId, const HistogramDataVector& data) {
-	printf("Histogram. ReqId: %d, data length: %lu\n", reqId, data.size());
+	printf("Histogram. ReqId: %d, data length: %zu\n", reqId, data.size());
 
 	for (auto item : data) {
 		printf("\t price: %f, size: %lld\n", item.price, item.size);
